fix stable version specific code
